pyproject.toml
[project]
name = "FinWell"
version = "0.1.0"
description = "AI Multi-Agent System for Financial Analysis & Investment Research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uv>=0.1.0",
    "crewai>=0.28.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "yfinance>=0.2.18",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "textblob>=0.17.1",
    "streamlit>=1.32.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.9.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["financial_ai_agents"]

[project.scripts]
financial-ai = "financial_ai_agents.main:main"
